requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2.1
django-cors-headers==4.1.0
django-extensions==3.2.1
django-jazzmin==2.6.0
django-location-field==2.7.1
django-nested-inline==0.4.6
django-ninja==0.21.0
django-silk==5.0.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
geographiclib==2.0
gprof2dot==2022.7.29
idna==3.4
Pillow==9.5.0
PyArabic==0.6.15
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.7
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
